New script to properly amend the test schema for Solr 5
The default test schema in Solr 5 doesn't have the fields used in the tests.
I wrote a small script to put the fields into the schema. It uses xml.etree
though, so I'm guessing it won't be Python 2.6 compliant... but how likely
is it that people clinging to Python 2.6 will be rushing toward Solr 5?